testing/web-platform/tests/battery-status/battery-promise.html

<!DOCTYPE html>
<title>navigator.getBattery() - navigator.getBattery()'s returnvalue is Promise<BatteryManager>. </title>
<link rel="author" title="YuichiNukiyama" href="https://github.com/YuichiNukiyama">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
    function returnBattery() {
        return navigator.getBattery();
    }

    promise_test(function () {
        return returnBattery()
          .then(function (result) {
              assert_class_string(result, "BatteryManager", "getBattery should return BatteryManager Object.");
          });
    }, "navigator.getBattery() return BatteryManager");

</script>

<div id="log"></div>
